Conflicts:
	App/EzServiceClient/Service References/EzServiceReference/Ezbob.Backend.Models.NewLoan.xsd
	App/EzServiceClient/Service References/EzServiceReference/Ezbob.Backend.ModelsWithDB.Experian.xsd
	App/EzServiceClient/Service References/EzServiceReference/Ezbob.Backend.ModelsWithDB.NewLoan.xsd
	App/EzServiceClient/Service References/EzServiceReference/Reference.cs
	App/EzServiceClient/Service References/EzServiceReference/Reference.svcmap
	App/EzServiceClient/Service References/EzServiceReference/configuration.svcinfo
	App/EzServiceClient/Service References/EzServiceReference/configuration91.svcinfo
	App/EzServiceClient/Service References/EzServiceReference/service.wsdl
	App/EzServiceClient/Service References/EzServiceReference/service.xsd
	App/EzServiceClient/app.config
	App/PluginWeb/EzBob.Web/Areas/Underwriter/Controllers/LoanEditorController.cs
	App/PluginWeb/EzBob.Web/Code/LoanCreator.cs
	Backend/Models/Alibaba/CustomerDataSharing.cs
	Backend/Models/NewLoan/ReschedulingArgument.cs
	Backend/Models/NewLoan/ReschedulingResult.cs
	Backend/ModelsWithDB/Ezbob.Backend.ModelsWithDB.csproj
	Backend/ModelsWithDB/NewLoan/NL_OfferFees.cs
	Backend/ModelsWithDB/NewLoan/NL_Offers.cs
	Backend/Strategies/MainStrategy/MainStrategy.cs
	Backend/Strategies/NewLoan/AddOffer.cs
	Backend/Strategies/Strategies.csproj
	EzBob.sln
	Integration/DbConstants/DbConstants.csproj
	Integration/DbConstants/NewLoanEnums.cs
	Integration/EzBob/EzBobTest/Misc.cs
	Integration/EzBob/EzBobTest/TestStrategies.cs
	sql/SPs/NL_BlendedLoansSave.sql.notinuseyet
	sql/SPs/NL_BlendedOffersSave.sql.notinuseyet
	sql/SPs/NL_CashRequestGetByOldID.sql
	sql/SPs/NL_CashRequestsSave.sql
	sql/SPs/NL_DecisionsSave.sql
	sql/SPs/NL_Exists.sql
	sql/SPs/NL_FundTransfersSave.sql
	sql/SPs/NL_LoanAgreementsSave.sql
	sql/SPs/NL_LoanFeePaymentsSave.sql
	sql/SPs/NL_LoanFeesSave.sql
	sql/SPs/NL_LoanHistorySave.sql
	sql/SPs/NL_LoanLegalsSave.sql
	sql/SPs/NL_LoanLienLinksSave.sql
	sql/SPs/NL_LoanOptionsSave.sql
	sql/SPs/NL_LoanRolloversSave.sql
	sql/SPs/NL_LoanSchedulePaymentsSave.sql
	sql/SPs/NL_LoanSchedulesSave.sql
	sql/SPs/NL_LoanStateLoad.sql.notinuse
	sql/SPs/NL_LoanStatesSave.sql
	sql/SPs/NL_LoansSave.sql
	sql/SPs/NL_OffersSave.sql
	sql/SPs/NL_PacnetTransactionsSave.sql
	sql/SPs/NL_PaymentsSave.sql
	sql/SPs/NL_PaypointTransactionsSave.sql
	sql/SPs/NL_RepaymentIntervalTypesLoad.sql
	sql/UsefullScripts/NL_drop.sql

+++ Backend/Strategies/AutoDecisionAutomation/AutoDecisions/Reject/Checker.cs
@@ -238,9 +238,9 @@
 
 		private void CompanyFiles() {
 			var data = new AutomationCalculator.ProcessHistory.AutoRejection.Turnover.DataModel {
-				AnnualTurnover = (int)Trail.MyInputData.AnnualTurnover,
+				AnnualTurnover = Trail.MyInputData.AnnualTurnover,
 				AnnualTurnoverThreshhold = Trail.MyInputData.TotalAnnualTurnover,
-				QuarterTurnover = (int)Trail.MyInputData.QuarterTurnover,
+				QuarterTurnover = Trail.MyInputData.QuarterTurnover,
 				QuarterTurnoverThreshhold = Trail.MyInputData.TotalThreeMonthTurnover,
 				HasCompanyFiles = Trail.MyInputData.HasCompanyFiles,
 			};
